Minimize initial capacity of the chunks deque
It should be normal for a buffer in typical usage, i.e. serializing
a message struct into a buffer chunked in the ballpark of expected size
of the serialization, to never have to grow to more than a few chunks.
If the output is fast, written data may be consumed directly from
the staging buffer without being splitting off to chunks.
It makes sense to only allocate the minimum capacity for the
chunks queue that the VecDeque will internally ensure anyway
(VecDeque::new() allocates to a larger capacity).